+++ piptools/repositories/pypi.py
@@ -14,19 +14,19 @@ from .._compat import (
     Link,
     PyPI,
     RequirementSet,
-    RequirementTracker,
     Resolver as PipResolver,
     TemporaryDirectory,
     Wheel,
     WheelCache,
     contextlib,
+    get_requirement_tracker,
+    global_tempdir_manager,
     is_dir_url,
     is_file_url,
     is_vcs_url,
     path_to_url,
     url_to_path,
 )
-from ..cache import CACHE_DIR
 from ..click import progressbar
 from ..exceptions import NoCandidateFound
 from ..logging import log
@@ -54,17 +54,17 @@ class PyPIRepository(BaseRepository):
     changed/configured on the Finder.
     """
 
-    def __init__(self, pip_args, build_isolation=False):
+    def __init__(self, pip_args, cache_dir, build_isolation=False):
         self.build_isolation = build_isolation
 
         # Use pip's parser for pip.conf management and defaults.
         # General options (find_links, index_url, extra_index_url, trusted_host,
         # and pre) are deferred to pip.
-        command = create_install_command()
-        self.options, _ = command.parse_args(pip_args)
+        self.command = create_install_command()
+        self.options, _ = self.command.parse_args(pip_args)
 
-        self.session = command._build_session(self.options)
-        self.finder = command._build_package_finder(
+        self.session = self.command._build_session(self.options)
+        self.finder = self.command._build_package_finder(
             options=self.options, session=self.session
         )
 
@@ -81,8 +81,9 @@ class PyPIRepository(BaseRepository):
 
         # Setup file paths
         self.freshen_build_caches()
-        self._download_dir = fs_str(os.path.join(CACHE_DIR, "pkgs"))
-        self._wheel_download_dir = fs_str(os.path.join(CACHE_DIR, "wheels"))
+        self._cache_dir = cache_dir
+        self._download_dir = fs_str(os.path.join(self._cache_dir, "pkgs"))
+        self._wheel_download_dir = fs_str(os.path.join(self._cache_dir, "wheels"))
 
     def freshen_build_caches(self):
         """
@@ -148,9 +149,14 @@ class PyPIRepository(BaseRepository):
             )
             best_candidate = best_candidate_result.best_candidate
 
+        if PIP_VERSION[:2] <= (19, 3):
+            best_candidate_name = best_candidate.project
+        else:
+            best_candidate_name = best_candidate.name
+
         # Turn the candidate into a pinned InstallRequirement
         return make_install_requirement(
-            best_candidate.project,
+            best_candidate_name,
             best_candidate.version,
             ireq.extras,
             constraint=ireq.constraint,
@@ -202,26 +208,45 @@ class PyPIRepository(BaseRepository):
                 )
                 resolver_kwargs["make_install_req"] = make_install_req
 
+            if PIP_VERSION >= (20,):
+                del resolver_kwargs["session"]
+                del preparer_kwargs["progress_bar"]
+
             resolver = None
             preparer = None
-            with RequirementTracker() as req_tracker:
+
+            if PIP_VERSION[:2] <= (19, 3):
+                tmp_dir_cm = contextlib.nullcontext()
+            else:
+                from pip._internal.utils.temp_dir import TempDirectory
+
+                tmp_dir_cm = TempDirectory(kind="req-tracker")
+
+            with get_requirement_tracker() as req_tracker, tmp_dir_cm as temp_build_dir:
                 # Pip 18 uses a requirement tracker to prevent fork bombs
                 if req_tracker:
                     preparer_kwargs["req_tracker"] = req_tracker
-                preparer = RequirementPreparer(**preparer_kwargs)
+
+                if PIP_VERSION[:2] <= (19, 3):
+                    preparer = RequirementPreparer(**preparer_kwargs)
+                else:
+                    preparer = self.command.make_requirement_preparer(
+                        temp_build_dir=temp_build_dir,
+                        options=self.options,
+                        req_tracker=req_tracker,
+                        session=self.session,
+                        finder=self.finder,
+                        use_user_site=self.options.use_user_site,
+                    )
+
                 resolver_kwargs["preparer"] = preparer
                 reqset = RequirementSet()
                 ireq.is_direct = True
                 reqset.add_requirement(ireq)
 
                 resolver = PipResolver(**resolver_kwargs)
-                require_hashes = False
-                if PIP_VERSION < (19, 4):
-                    resolver.require_hashes = require_hashes
-                    results = resolver._resolve_one(reqset, ireq)
-                else:  # pragma: no cover
-                    # TODO remove pragma after pip==19.4 being released
-                    results = resolver._resolve_one(reqset, ireq, require_hashes)
+                resolver.require_hashes = False
+                results = resolver._resolve_one(reqset, ireq)
 
                 reqset.cleanup_files()
 
@@ -259,22 +284,23 @@ class PyPIRepository(BaseRepository):
             if not os.path.isdir(self._wheel_download_dir):
                 os.makedirs(self._wheel_download_dir)
 
-            wheel_cache = WheelCache(CACHE_DIR, self.options.format_control)
+            wheel_cache = WheelCache(self._cache_dir, self.options.format_control)
             prev_tracker = os.environ.get("PIP_REQ_TRACKER")
             try:
-                self._dependencies_cache[ireq] = self.resolve_reqs(
-                    download_dir, ireq, wheel_cache
-                )
+                with global_tempdir_manager():
+                    self._dependencies_cache[ireq] = self.resolve_reqs(
+                        download_dir, ireq, wheel_cache
+                    )
             finally:
                 if "PIP_REQ_TRACKER" in os.environ:
                     if prev_tracker:
                         os.environ["PIP_REQ_TRACKER"] = prev_tracker
                     else:
                         del os.environ["PIP_REQ_TRACKER"]
-                try:
-                    self.wheel_cache.cleanup()
-                except AttributeError:
-                    pass
+
+                # WheelCache.cleanup() introduced in pip==10.0.0
+                if PIP_VERSION >= (10,):
+                    wheel_cache.cleanup()
         return self._dependencies_cache[ireq]
 
     def get_hashes(self, ireq):